+++ apps/backoffice-v2/src/lib/blocks/hooks/useUbosBlock/useUbosBlock.tsx
@@ -1,9 +1,74 @@
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
+import { WarningFilledSvg } from '@/common/components/atoms/icons';
 
-export const useUbosBlock = ubos => {
+type Ubo = {
+  name?: string;
+  type?: string;
+  level?: number;
+  percentage?: number;
+};
+
+export const useUbosBlock = (ubos: Ubo[] | undefined, message: string | undefined) => {
+  const getCell = useCallback(() => {
+    if (Array.isArray(ubos) && ubos?.length) {
+      return {
+        type: 'table',
+        value: {
+          columns: [
+            {
+              accessorKey: 'name',
+              header: 'Name',
+            },
+            {
+              accessorKey: 'percentage',
+              header: 'Percentage (25% or higher)',
+            },
+            {
+              accessorKey: 'type',
+              header: 'Type',
+            },
+            {
+              accessorKey: 'level',
+              header: 'Level',
+            },
+          ],
+          data: ubos,
+        },
+      } satisfies Extract<
+        Parameters<ReturnType<typeof createBlocksTyped>['addCell']>[0],
+        {
+          type: 'table';
+        }
+      >;
+    }
+
+    if (message) {
+      return {
+        type: 'paragraph',
+        value: (
+          <span className="flex text-sm text-black/60">
+            <WarningFilledSvg
+              className={'mr-[8px] mt-px text-black/20'}
+              width={'20'}
+              height={'20'}
+            />
+            <span>{message}</span>
+          </span>
+        ),
+      } satisfies Extract<
+        Parameters<ReturnType<typeof createBlocksTyped>['addCell']>[0],
+        {
+          type: 'paragraph';
+        }
+      >;
+    }
+  }, [ubos]);
+
   return useMemo(() => {
-    if (!Array.isArray(ubos) || !ubos?.length) {
+    const cell = getCell();
+
+    if (!cell) {
       return [];
     }
 
@@ -24,38 +89,10 @@ export const useUbosBlock = ubos => {
               className: 'mb-4',
             },
           })
-          .addCell({
-            type: 'table',
-            value: {
-              columns: [
-                {
-                  accessorKey: 'name',
-                  header: 'Name',
-                },
-                {
-                  accessorKey: 'percentage',
-                  header: 'Percentage (25% or higher)',
-                },
-                {
-                  accessorKey: 'type',
-                  header: 'Type',
-                },
-                {
-                  accessorKey: 'level',
-                  header: 'Level',
-                },
-              ],
-              data: ubos.map(({ name, percentage, type, level }) => ({
-                name,
-                percentage,
-                type,
-                level,
-              })),
-            },
-          })
+          .addCell(cell)
           .build()
           .flat(1),
       })
       .build();
-  }, [ubos]);
+  }, [getCell]);
 };