+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useFilterId } from '@/common/hooks/useFilterId/useFilterId';
-import { useWorkflowQuery } from '@/domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
+import { useWorkflowByIdQuery } from '@/domains/workflows/hooks/queries/useWorkflowByIdQuery/useWorkflowByIdQuery';
 import { useAuthenticatedUserQuery } from '@/domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
 import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
 import { useCaseDecision } from '@/pages/Entity/components/Case/hooks/useCaseDecision/useCaseDecision';
@@ -41,10 +41,18 @@ import {
   useAssociatedCompaniesBlock,
 } from '@/lib/blocks/hooks/useAssosciatedCompaniesBlock/useAssociatedCompaniesBlock';
 
+const pluginsOutputBlacklist = [
+  'companySanctions',
+  'directors',
+  'ubo',
+  'businessInformation',
+  'website_monitoring',
+] as const;
+
 export const useDefaultBlocksLogic = () => {
   const { entityId: workflowId } = useParams();
   const filterId = useFilterId();
-  const { data: workflow, isLoading } = useWorkflowQuery({
+  const { data: workflow, isLoading } = useWorkflowByIdQuery({
     workflowId,
     filterId,
   });
@@ -123,17 +131,12 @@ export const useDefaultBlocksLogic = () => {
   const kybChildWorkflows = workflow?.childWorkflows?.filter(
     childWorkflow => childWorkflow?.context?.entity?.type === 'business',
   );
-  const pluginsOutputBlacklist = [
-    'companySanctions',
-    'directors',
-    'ubo',
-    'businessInformation',
-    'website_monitoring',
-  ] as const;
+
   const filteredPluginsOutput = useMemo(
     () => omitPropsFromObject(workflow?.context?.pluginsOutput, ...pluginsOutputBlacklist),
     [pluginsOutputBlacklist, workflow?.context?.pluginsOutput],
   );
+
   const pluginsOutputKeys = Object.keys(filteredPluginsOutput ?? {});
   const directorsDocuments = useMemo(() => selectDirectorsDocuments(workflow), [workflow]);
   const directorDocumentPages = useMemo(
@@ -143,6 +146,7 @@ export const useDefaultBlocksLogic = () => {
       ),
     [directorsDocuments],
   );
+
   const directorsStorageFilesQueryResult = useStorageFilesQuery(directorDocumentPages);
   const directorsDocumentPagesResults: Array<Array<string>> = useDocumentPageImages(
     directorsDocuments,
@@ -164,12 +168,14 @@ export const useDefaultBlocksLogic = () => {
       places: sanction?.entity?.places,
     }),
   );
+
   const ubos = workflow?.context?.pluginsOutput?.ubo?.data?.uboGraph?.map(ubo => ({
     name: ubo?.name,
     percentage: ubo?.shareHolders?.[0]?.sharePercentage,
     type: ubo?.type,
     level: ubo?.level,
   }));
+
   const directorsRegistryProvided = workflow?.context?.pluginsOutput?.directors?.data?.map(
     ({ name, position }) => ({
       name,
@@ -269,7 +275,7 @@ export const useDefaultBlocksLogic = () => {
 
   const companySanctionsBlock = useCompanySanctionsBlock(companySanctions);
 
-  const ubosBlock = useUbosBlock(ubos);
+  const ubosBlock = useUbosBlock(ubos, workflow?.context?.pluginsOutput?.ubo?.message);
 
   const directorsUserProvidedBlock = useDirectorsUserProvidedBlock(directorsUserProvided);