+++ packages/react-pdf-toolkit/src/components/atoms/List/List.Item.tsx
@@ -1,0 +1,16 @@
+import { ListItemProps } from '@/components/atoms/List/types';
+import { tw } from '@/theme';
+import { mergeStyles } from '@/utils/merge-styles';
+import { View } from '@react-pdf/renderer';
+import { FunctionComponent } from 'react';
+
+export const ListItem: FunctionComponent<ListItemProps> = ({
+  children,
+  prependIcon,
+  styles = [],
+}) => (
+  <View style={mergeStyles([tw('flex flex-row items-center gap-3'), ...styles])}>
+    {prependIcon && <View>{prependIcon}</View>}
+    <View>{children}</View>
+  </View>
+);