+++ services/workflows-service/scripts/seed.ts
@@ -31,6 +31,8 @@ import {
   baseFilterEndUserSelect,
 } from './filters';
 import { generateTransactions } from './workflows/generate-transactions';
+import { generateKycManualReviewRuntimeAndToken } from './workflows/runtime/geneate-kyc-manual-review-runtime-and-token';
+import { Type } from '@sinclair/typebox';
 
 seed(10).catch(error => {
   console.error(error);
@@ -70,7 +72,7 @@ async function createCustomer(
   faviconImageUri: string,
   webhookSharedSecret: string,
 ) {
-  return await client.customer.create({
+  return client.customer.create({
     data: {
       id: `customer-${id}`,
       name: `customer-${id}`,
@@ -102,7 +104,10 @@ async function createProject(client: PrismaClient, customer: Customer, id: strin
 
 const DEFAULT_INITIAL_STATE = CommonWorkflowStates.MANUAL_REVIEW;
 
-const DEFAULT_SEED_DEFINITION_TOKEN = '12345678-1234-1234-1234-123456789012';
+const DEFAULT_TOKENS = {
+  KYB: '12345678-1234-1234-1234-123456789012',
+  KYC: '12345678-1234-1234-1234-123456789000',
+};
 
 async function seed(bcryptSalt: string | number) {
   console.info('Seeding database...');
@@ -186,6 +191,8 @@ async function seed(bcryptSalt: string | number) {
   const kycManualMachineId = 'MANUAL_REVIEW_0002zpeid7bq9aaa';
   const kybManualMachineId = 'MANUAL_REVIEW_0002zpeid7bq9bbb';
   const manualMachineVersion = 1;
+
+  const kycWorkflowDefinitionId = 'kyc-manual-review';
 
   const onboardingMachineKycId = 'COLLECT_DOCS_b0002zpeid7bq9aaa';
   const onboardingMachineKybId = 'COLLECT_DOCS_b0002zpeid7bq9bbb';
@@ -636,6 +643,54 @@ async function seed(bcryptSalt: string | number) {
     },
   });
 
+  const getDocumentsSchema = () =>
+    ['id_card', 'passport', 'drivers_license', 'voter_id'].map(name => ({
+      category: name,
+      type: name,
+      issuer: { country: 'ZZ' },
+      issuingVersion: 1,
+      version: 1,
+      propertiesSchema: Type.Object({
+        firstName: Type.Optional(Type.String()),
+        lastName: Type.Optional(Type.String()),
+        documentNumber: Type.Optional(Type.String()),
+        dateOfBirth: Type.Optional(Type.String({ format: 'date' })),
+        expirationDate: Type.Optional(Type.String({ format: 'date' })),
+        isFaceMatching: Type.Optional(Type.Boolean()),
+      }),
+    }));
+
+  await client.workflowDefinition.create({
+    data: {
+      ...baseReviewDefinition(generateBaseTaskLevelStates()),
+      id: kycWorkflowDefinitionId,
+      documentsSchema: getDocumentsSchema(),
+      config: {
+        workflowLevelResolution: false,
+        availableDocuments: [
+          {
+            category: 'id_card',
+            type: 'id_card',
+          },
+          {
+            category: 'passport',
+            type: 'passport',
+          },
+          {
+            category: 'drivers_license',
+            type: 'drivers_license',
+          },
+          {
+            category: 'voter_id',
+            type: 'voter_id',
+          },
+        ],
+      },
+      version: 3,
+      projectId: project1.id,
+    },
+  });
+
   // KYB
   await client.workflowDefinition.create({
     data: {
@@ -777,6 +832,30 @@ async function seed(bcryptSalt: string | number) {
     project1.id,
   );
 
+  await createFilter(
+    'KYC - Manual Review',
+    'individuals',
+    {
+      select: {
+        id: true,
+        status: true,
+        assigneeId: true,
+        context: true,
+        createdAt: true,
+        state: true,
+        tags: true,
+        ...baseFilterDefinitionSelect,
+        ...baseFilterEndUserSelect,
+        ...baseFilterAssigneeSelect,
+      },
+      where: {
+        workflowDefinitionId: { in: [kycWorkflowDefinitionId] },
+        endUserId: { not: null },
+      },
+    },
+    project1.id,
+  );
+
   // KYB Onboarding
   await client.workflowDefinition.create({
     data: {
@@ -891,8 +970,7 @@ async function seed(bcryptSalt: string | number) {
         childWorkflowsRuntimeData: true,
       },
       where: {
-        // workflowDefinitionId: 'dynamic_kyb_parent_example',
-        workflowDefinitionId: 'kyb_with_associated_companies_example',
+        workflowDefinitionId: { in: ['kyb_with_associated_companies_example'] },
         businessId: { not: null },
         state: {
           in: [
@@ -1011,7 +1089,14 @@ async function seed(bcryptSalt: string | number) {
     projectId: project1.id,
     endUserId: endUserIds[0]!,
     businessId: businessIds[0]!,
-    token: DEFAULT_SEED_DEFINITION_TOKEN,
+    token: DEFAULT_TOKENS.KYB,
+  });
+
+  await generateKycManualReviewRuntimeAndToken(client, {
+    workflowDefinitionId: kycWorkflowDefinitionId,
+    projectId: project1.id,
+    endUserId: endUserIds[0]!,
+    token: DEFAULT_TOKENS.KYC,
   });
 
   console.info('Seeded database successfully');