+++ services/workflows-service/src/common/access-control/abac.util.ts
@@ -4,10 +4,10 @@ import { Permission } from 'accesscontrol';
  * @returns attributes not allowed to appear on given data according to given
  * attributeMatchers
  */
-export function getInvalidAttributes(
+export const getInvalidAttributes = (
   permission: Permission,
   data: Record<string, unknown>,
-): string[] {
+): string[] => {
   const filteredData = permission.filter(data) as Record<string, unknown>;
   return Object.keys(data).filter(key => !(key in filteredData));
-}
+};