+++ services/workflows-service/src/workflow/schemas/zod-schemas.ts
@@ -12,18 +12,20 @@ export const SubscriptionSchema = z.discriminatedUnion('type', [
 
 export const ConfigSchema = z
   .object({
+    isCaseOverviewEnabled: z.boolean().optional(),
+    isLegacyReject: z.boolean().optional(),
+    isLockedDocumentCategoryAndType: z.boolean().optional(),
+    isManualCreation: z.boolean().optional(),
+    isDemo: z.boolean().optional(),
+    isExample: z.boolean().optional(), // OSS only
     language: z.string().optional(),
     supportedLanguages: z.array(z.string()).optional(),
-    isLegacyReject: z.boolean().optional(),
     subscriptions: z.array(SubscriptionSchema).optional(),
     completedWhenTasksResolved: z.boolean().optional(),
     workflowLevelResolution: z.boolean().optional(),
-    isLockedDocumentCategoryAndType: z.boolean().optional(),
     allowMultipleActiveWorkflows: z.boolean().optional(),
     initialEvent: z.string().optional(),
     availableDocuments: z.array(z.object({ category: z.string(), type: z.string() })).optional(),
-    isDemo: z.boolean().optional(),
-    isExample: z.boolean().optional(), // OSS only
     callbackResult: z
       .object({
         transformers: z.array(z.any()),