+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/report-generation-example
 
+## 0.0.2
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.0.12
+
 ## 0.0.1
 
 ### Patch Changes