+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # web-ui-sdk
 
+## 1.4.37
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.40
+
+## 1.4.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.39
+
+## 1.4.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.38
+
 ## 1.4.34
 
 ### Patch Changes